🦄 refactor(Refactor Types and Run Linting):
Changes to the following areas:
 - Removing paths from the tsconfig. These break with nodemon running.
 - Refactoring handling utils and types to match changes in ServerInit.
 - Expirementing with the linter and linting the project.
