+++ .eslintrc.json
@@ -49,7 +49,8 @@
     "rules": {
       // General JavaScript/TypeScript rules
       "no-console": "warn",
-      "no-unused-vars": "error",
+      "no-unused-vars": "off",
+      "@typescript-eslint/no-unused-vars": "error",
       "prefer-rest-params": "error",
       "@typescript-eslint/consistent-type-imports": "error",
       "@typescript-eslint/no-floating-promises": "error",
