+++ package.json
@@ -8,7 +8,7 @@
     "dev": "nodemon server/server.ts",
     "build": "next build",
     "start": "next start",
-    "lint": "eslint . --quiet && prettier --write ."
+    "lint": "eslint . --fix"
   },
   "keywords": [
     "education"
