+++ server/routes/genAi.ts
@@ -1,22 +1,22 @@
-import express from "express";
-import type { Request, Response } from "express";
-import genAiController from "../controllers/genAiController";
+import express from 'express';
+import type { Request, Response } from 'express';
+import genAiController from '@/server/controllers/genAiController';
 
 const router = express.Router();
 
 // default route for testing. This route should be removed in production.
-router.get("/", (_: Request, res: Response) => {
-  res.send({ message: "Hello from the AI!" });
+router.get('/', (_: Request, res: Response) => {
+  res.send({ message: 'Hello from the AI!' });
 });
 
 /**
  * Ask a single question of the AI.
  */
 router.post(
-  "/ask",
+  '/ask',
   genAiController.generateResponse,
   (_: Request, res: Response) => {
-    res.send({ message: "AI response", response: res.locals.response });
+    res.send({ message: 'AI response', response: res.locals.response });
   }
 );
