+++ server/types/ai.d.ts
@@ -1,0 +1,19 @@
+import type { GoogleGenerativeAI, InputContent } from "@google/generative-ai";
+
+export interface AiInterface {
+  apiKey: string;
+
+  // ai will need more types if we expand the factory to include other AI models
+  ai: GoogleGenerativeAI;
+  connect(): void;
+
+  // these methods are based upon the Gemini model, they will need to be updated if we add more models
+  generateResponse(prompt: string): Promise<string>;
+  generateConversation(
+    history: InputContent[],
+    prompt: string,
+    generationConfig: object
+  ): Promise<string>;
+}
+
+export type AiTypes = "gemini";
